# core, core/state, trie: enterprise hand-tuned multi-level caching