@karalabe 2016-05-26
@obscuren @fjl Squashed and ready for final nod/merge.

Note: I retained this copy https://github.com/ethereum/go-ethereum/pull/2585/files#diff-18ac19c82220cf468504a04b97b4c584R538 as without it the iterator tests break. Which means, that seemingly we have some side-effect somewhere that breaks the short node keys.